move convert_df_to_time_intervals to tools.text
